Modified the application to utilize the ESPA internal file format.  Modified thermal band usage to no longer convert to Kelvin, due to the recent change of lndcal/lndth to write thermal values in Kelvin.

+++ trunk/ledapsSrc/src/lndsr/clouds.h
@@ -26,12 +26,9 @@ int free_cld_diags(struct cld_diags_t *cld_diags);
 void fill_cld_diags(cld_diags_t *cld_diags);
 int interpol_clddiags_1pixel(cld_diags_t *cld_diags, int img_line, int img_sample,float *inter_value);
 
-bool cloud_detection_pass1(Lut_t *lut, int nsamp, int il, int **line_in, int8 *qa_line, int *b6_line,float *atemp_line,
-		  cld_diags_t *cld_diags);
-bool cloud_detection_pass2(Lut_t *lut, int nsamp, int il, int **line_in, int8 *qa_line, int *b6_line,
-		  cld_diags_t *cld_diags,char *ddv_line);
-bool cast_cloud_shadow(Lut_t *lut, int nsamp, int il_start, int ***line_in, int **b6_line,
-		  cld_diags_t *cld_diags,char ***cloud_buf, Ar_gridcell_t *ar_gridcell,float pixel_size,float adjust_north);
+bool cloud_detection_pass1(Lut_t *lut, int nsamp, int il, int16 **line_in, uint8 *qa_line, int16 *b6_line,float *atemp_line, cld_diags_t *cld_diags);
+bool cloud_detection_pass2(Lut_t *lut, int nsamp, int il, int16 **line_in, uint8 *qa_line, int16 *b6_line, cld_diags_t *cld_diags,char *ddv_line);
+bool cast_cloud_shadow(Lut_t *lut, int nsamp, int il_start, int16 ***line_in, int16 **b6_line, cld_diags_t *cld_diags, char ***cloud_buf, Ar_gridcell_t *ar_gridcell, float pixel_size, float adjust_north);
 bool dilate_cloud_mask(Lut_t *lut, int nsamp, char ***cloud_buf, int dilate_dist);
 bool dilate_shadow_mask(Lut_t *lut, int nsamp, char ***cloud_buf, int dilate_dist);
